refactor: url 패턴 변경
