+++ src/test/java/com/example/board/domain/member/controller/v1/MemberControllerTest.java
@@ -63,7 +63,7 @@ class MemberControllerTest {
     MemberRequest.SignUp signUpRequest = new MemberRequest.SignUp("김환", "email123@naver.com", "password123!", 25, "게임");
 
     //when & then
-    mockMvc.perform(post("/api/v1/members")
+    mockMvc.perform(post("/members/v1")
             .contentType(MediaType.APPLICATION_JSON)
             .content(objectMapper.writeValueAsString(signUpRequest)))
         .andExpect(status().isCreated())
@@ -89,7 +89,7 @@ class MemberControllerTest {
     MemberRequest.Login loginMember = new MemberRequest.Login("email123@naver.com", "password123!");
 
     //when & then
-    mockMvc.perform(post("/api/v1/members/login")
+    mockMvc.perform(post("/members/v1/login")
             .contentType(MediaType.APPLICATION_JSON)
             .content(objectMapper.writeValueAsString(loginMember)))
         .andExpect(status().isOk())
@@ -111,7 +111,7 @@ class MemberControllerTest {
     postService.save(new PostRequest("으앙", "응아아아아앙", savedMemberId));
 
     //when & then
-    mockMvc.perform(get("/api/v1/members/mypage")
+    mockMvc.perform(get("/members/v1/mypage")
             .cookie(new Cookie("loginId", savedMemberId.toString())))
         .andExpect(status().isOk())
         .andDo(print())
@@ -134,7 +134,7 @@ class MemberControllerTest {
   @Test
   @DisplayName("로그아웃할 수 있습니다.")
   void logout() throws Exception {
-    mockMvc.perform(post("/api/v1/members/logout"))
+    mockMvc.perform(post("/members/v1/logout"))
         .andExpect(status().isOk())
         .andDo(print())
         .andDo(document("member-logout"));
